배포

+++ payment/src/main/java/com/example/payment/application/dto/CreateWalletResult.java
@@ -1,11 +1,12 @@
 package com.example.payment.application.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record CreateWalletResult(
         UUID walletId,
         UUID memberId,
-        Long balance,
+        BigDecimal balance,
         boolean created
 ) {
 }
