+++ payment/src/main/java/com/example/payment/application/dto/OrderPaymentLineCommand.java
@@ -1,5 +1,6 @@
 package com.example.payment.application.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -8,6 +9,6 @@ import java.util.UUID;
 public record OrderPaymentLineCommand(
         UUID orderItemId,
         UUID sellerMemberId,
-        Long lineAmount
+        BigDecimal lineAmount
 ) {
 }
