+++ payment/src/main/java/com/example/payment/application/dto/PaymentRefundItemCommand.java
@@ -1,9 +1,10 @@
 package com.example.payment.application.dto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record PaymentRefundItemCommand(
         UUID orderItemId,
-        Long refundAmount
+        BigDecimal refundAmount
 ) {
 }
