+++ payment/src/main/java/com/example/payment/application/dto/WalletTransactionItemResult.java
@@ -1,6 +1,7 @@
 package com.example.payment.application.dto;
 
 import com.example.payment.domain.enumtype.WalletTransactionType;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -10,8 +11,8 @@ import java.util.UUID;
 public record WalletTransactionItemResult(
         UUID transactionId,
         WalletTransactionType transactionType,
-        Long amount,
-        Long balanceAfter,
+        BigDecimal amount,
+        BigDecimal balanceAfter,
         String referenceType,
         UUID referenceId,
         String description,
