+++ payment/src/main/java/com/example/payment/infrastructure/messaging/kafka/contract/SettlementCandidateCreatedMessage.java
@@ -1,6 +1,7 @@
 package com.example.payment.infrastructure.messaging.kafka.contract;
 
 import com.example.payment.domain.enumtype.ConfirmationType;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -9,7 +10,7 @@ public record SettlementCandidateCreatedMessage(
         UUID orderId,
         UUID escrowId,
         UUID sellerMemberId,
-        Long grossAmount,
+        BigDecimal grossAmount,
         Instant releasedAt,
         ConfirmationType confirmationType,
         Instant occurredAt
