+++ settlement/src/main/java/com/example/settlement/application/dto/SettlementItemCreateCommand.java
@@ -1,5 +1,6 @@
 package com.example.settlement.application.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -7,7 +8,7 @@ public record SettlementItemCreateCommand(
         UUID orderId,
         UUID escrowId,
         UUID sellerId,
-        Long grossAmount,
+        BigDecimal grossAmount,
         LocalDateTime releasedAt
 ) {
 }
